+++ packages/dag-nodes/llm-text-openai/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @robota-sdk/dag-node-llm-text-openai
 
+## 3.0.0-beta.54
+
+### Patch Changes
+
+- Updated dependencies
+  - @robota-sdk/agent-core@3.0.0-beta.54
+  - @robota-sdk/agent-provider-openai@3.0.0-beta.54
+  - @robota-sdk/dag-core@3.0.0-beta.54
+  - @robota-sdk/dag-node@3.0.0-beta.54
+
 ## 3.0.0-beta.53
 
 ### Patch Changes